+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [v1.1.136] - 2024-05-22
+
+### Changed
+
+- Updated [@adguard/filters-downloader] to v2.2.1
+
+[v1.1.135]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.135...v1.1.136
+
+
 ## [v1.1.135] - 2024-05-20
 
 ### Added